requirements.txt
# Core Testing Dependencies
pytest==7.4.3
pytest-xdist==3.5.0
pytest-html==4.1.1
pytest-timeout==2.2.0

# Selenium WebDriver
selenium==4.16.0
webdriver-manager==4.0.1

# Percy Visual Testing
percy-selenium==2.0.2

# Page Object Support
selenium-page-factory==1.1.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0

# Reporting
allure-pytest==2.13.2

# Linting & Formatting (dev)
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
pylint==3.0.3
